Update train code
