+++ bts/swinunetr/common.py
@@ -1,0 +1,30 @@
+import argparse
+
+
+def get_args() -> argparse.Namespace:
+    """Retrieve arguments passed to the script.
+
+    Returns:
+        Namespace object where each argument can be accessed using the dot notation.
+    """
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--data-dir",
+        type=str,
+        required=True,
+        help="Path to dataset directory.",
+    )
+    parser.add_argument(
+        "--hyperparameters",
+        type=str,
+        help=(
+            "Training hyperparameters file path. The file in the path given should be "
+            "a valid yaml file. If not specified script will look for a "
+            "``hyperparameters.yaml`` file in the same directory the script is located "
+            "in."
+        ),
+    )
+    parser.add_argument(
+        "--output", type=str, required=True, help="Path to save the trained model."
+    )
+    return parser.parse_args()
