+++ bts/swinunetr/hyperparameters.yaml
@@ -1,29 +1,13 @@
-BATCH_SIZE: 2
+BATCH_SIZE: 3
 
-SHUFFLE: true
-
-ROI: [128, 128, 128]
+IMAGE_SIZE: [128, 128, 128]
 
 IN_CHANNELS: 1
 OUT_CHANNELS: 2
 
-FEATURE_SIZE: 48
 
-GRADIENT_CHECKPOINT: TRUE
-
-EPOCHS: 1000
+EPOCHS: 5000
 FOLD: 1
-
-# Batch size to run window slices
-SW_BATCH_SIZE: 2
-
-# 1e-4
-LEARNING_RATE: 0.0001
-
-INFER_OVERLAP: 0.6
-
-# 1e-5
-WEIGHT_DECAY: 0.00001
 
 DEVICE: cuda
